requirements.txt
fastapi==0.115.9
icalendar==6.1.1
langchain_core==0.3.40
langchain_openai==0.3.7
pydantic==2.10.6
python-dotenv==1.0.1
slowapi==0.1.9
uvicorn==0.34.0
